requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
python-dotenv
-e .